refactor to datex expression traits
