+++ crates/datex-core/src/shared_values/traits/mod.rs
@@ -1,4 +1,8 @@
 mod shared_container_common;
 pub use shared_container_common::*;
-mod expose_rc_internal; // IMPORTANT: don't expose this module, for internal use only
+mod expose_rc_internal;
+#[cfg(feature = "decompiler")]
+mod to_datex_expression_data;
+
+// IMPORTANT: don't expose this module, for internal use only
 pub(crate) use expose_rc_internal::_ExposeRcInternal;
