+++ crates/datex-core/src/types/shared_container_containing_entity_type/mod.rs
@@ -1,3 +1,6 @@
+#[cfg(feature = "decompiler")]
+mod to_type_expression_data;
+
 use crate::{
     prelude::*,
     runtime::pointer_address_provider::SelfOwnedPointerAddressProvider,
