+++ crates/datex-core/src/types/type/mod.rs
@@ -343,6 +343,8 @@ impl Type {
 
 pub mod equality;
 pub mod serde_dif;
+#[cfg(feature = "decompiler")]
+mod to_type_expression_data;
 
 impl Display for Type {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
