+++ crates/datex-core/src/types/type/to_type_expression_data.rs
@@ -1,0 +1,16 @@
+use crate::ast::type_expressions::TypeExpressionData;
+use crate::traits::to_type_expression_data::ToTypeExpressionData;
+use crate::types::r#type::Type;
+
+impl ToTypeExpressionData for Type {
+    fn to_type_expression_data(&self) -> TypeExpressionData {
+        match self {
+            Type::Entity(container) => {
+                container.to_type_expression_data()
+            }
+            Type::Definition(definition) => {
+                definition.to_type_expression_data()
+            }
+        }
+    }
+}
